+++ packages/terra-data-grid/src/DataGrid.jsx
@@ -114,6 +114,14 @@ const propTypes = {
   onRangeSelection: PropTypes.func,
 
   /**
+   * Callback function that is called when a cell range selection occurs. Parameters:
+   * @param {number} rowIndex RowIndex of the cell from which the range selection was triggered.
+   * @param {number} columnIndex ColumnIndex of the cell from which the range selection was triggered.
+   * @param {number} direction Direction keycode representing the direction of the selection.
+   */
+  onCellRangeSelect: PropTypes.func,
+
+  /**
    * Boolean indicating whether or not the DataGrid should allow entire rows to be selectable. An additional column will be
    * rendered to allow for row selection to occur.
    */
@@ -147,6 +155,7 @@ const DataGrid = injectIntl((props) => {
     onCellSelect,
     onClearSelection,
     onRangeSelection,
+    onCellRangeSelect,
     hasSelectableRows,
     rowHeaderIndex,
   } = props;
@@ -495,8 +504,13 @@ const DataGrid = injectIntl((props) => {
         return;
     }
 
-    if (onRangeSelection && event.shiftKey && (event.keyCode === KeyCode.KEY_UP || event.keyCode === KeyCode.KEY_DOWN)) {
+    const upDownKeys = new Set([KeyCode.KEY_UP, KeyCode.KEY_DOWN]);
+    const directionalKeys = new Set([KeyCode.KEY_UP, KeyCode.KEY_DOWN, KeyCode.KEY_LEFT, KeyCode.KEY_RIGHT]);
+    if (onRangeSelection && event.shiftKey && upDownKeys.has(event.keyCode)) {
       onRangeSelection(cellCoordinates.row, cellCoordinates.col, event.keyCode);
+    }
+    if (onCellRangeSelect && event.shiftKey && directionalKeys.has(event.keyCode)) {
+      onCellRangeSelect(cellCoordinates.row, cellCoordinates.col, event.keyCode);
     }
 
     if (nextRow > rows.length || nextCol >= displayedColumns.length) {