+++ packages/terra-data-grid/CHANGELOG.md
@@ -7,9 +7,12 @@
   * Fixed pinned column divider issue during keyboard resize.
   * Fixed resize handle being communicated when non-header cell is focused.
 
+* Added
+  * Added ability to select a range of cells to FlowsheetDataGrid.
+
 ## 1.0.0 - (October 25, 2023)
 * Initial working component release.
-  
+
 * Added
   * Added additional screen reader support to announce that column headers are interactable upon selection.
   * Added keyboard support for column resizing.